requirement.txt
langchain==0.2.1
langchain-google-genai==1.0.1
google-generativeai==0.3.2
faiss-cpu==1.7.4
pypdf==3.17.4
python-dotenv==1.0.1
